+++ application/views/itemadd.php
@@ -1,1 +1,1 @@
-<a href="/mtce/add"><input type="button" value="Add a new item"/></a>
+<a href="/mtceFleet/add"><input type="button" value="Add a new item"/></a>