+++ application/views/oneitemFlight.php
@@ -1,0 +1,8 @@
+<tr>
+	<td>{flightID}</td>
+	<td>{vehicleID}</td>
+	<td>{departureTime}</td>
+	<td>{arrivalTime}</td>
+	<td>{base}</td>
+	<td>{dest}</td>
+</tr>